refactored packet names; added toString()
