+++ src/main/java/scadinspect/data/scaddoc/properties/impl/IntSingleProperty.java
@@ -1,11 +1,14 @@
-package scadinspect.data.scaddoc.Properties;
+package scadinspect.data.scaddoc.properties.impl;
+
+import scadinspect.data.scaddoc.properties.SingleProperty;
 
 /**
  * Created by richteto on 17.03.2017.
  */
-public class IntSingleProperty implements Property {
-private String key;
-private long value;
+public class IntSingleProperty extends SingleProperty {
+
+  private String key;
+  private long value;
 
   public IntSingleProperty(String key, long value) {
     this.key = key;
